updates #4

+++ database/migrations/2023_06_18_232512_create_categories_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
 
             $table->string('title');
 
-            $table->integer('weight')->unsigned()->nullable();
+            $table->decimal('weight')->unsigned()->nullable();
             $table->integer('include_in')->unsigned();
 
             $table->boolean('active')->default(true);
